+++ server/src/FootballSubscriber.Core/Interfaces/IApiEntity.cs
@@ -5,5 +5,5 @@ public interface IApiEntity
     /// <summary>
     ///     Third party unique identifier
     /// </summary>
-    int ApiId { get; set; }
+    long ApiId { get; set; }
 }
